==> Scripts/Data_analysis/2_Descr_stats_BL_Paper1.Rmd <==
---
title: "Descr_BPD_memo"
output: html_document
---


```{r get_functions, eval=T, echo=F}
source("../functions_bpd_memo.R")

# load packages
packages <- c("ggplot2", "tidyverse", "lubridate", "data.table", "psych")
load_my_packages(packages)

# user-set options
overwrite_data = FALSE  # if TRUE, then the final data will be saved (overwritten)

```

```{r get_format_data, eval=T, echo=F}
data_final_bl <- read.csv2("../../Data/data_final_bl_180910.csv") # where all data were merged (baseline-only), and clean (attention check)

# set formats
cols_to_factor <- c("gender.sc", "gender.t1", "edu_years_compl.t1", "edu_highest.t1", "SES_subj.t1", "national.t1", 
                    "work.t1", "study.t1", "relationship.t1", "relationship_qual.t1", 
                    "English_first.sc", "English_level.sc", "English_first.t1", "English_level.t1", 
                    "medication.t1", "diagn_bpd.t1")
for (i in cols_to_factor) {
  data_final_bl[, i] <- as.factor(data_final_bl[, i]) 
}

levels(data_final_bl$gender.sc) <- c("1"="female", "2"="male", "3"="other", "4"="I don't want to tell")
levels(data_final_bl$gender.t1) <- c("1"="female", "2"="male", "3"="other", "4"="I don't want to tell")
levels(data_final_bl$edu_highest.t1) <- c("0"="No formal education", "1"="Primary school (or equivalent)", 
                                          "2"="Secondary/high school (or equivalent)", "3"="Vocational school (or equivalent)", 
                                          "4"="Undergraduate school/BA degree (or equivalent)", "5"="Graduate school/MA degree (or equivalent)", 
                                          "6"="PhD/MBA or similar")
levels(data_final_bl$edu_years_compl.t1) <- c("1" = "4 years or less",	"2" =	"5-8 years", 	"3" =	"9-12 years", 
                                              "4" = "12-17 years", "5" = "17-20 years", "6" =	"more than 20 years")
levels(data_final_bl$edu_highest.t1) <- c("0" = "No formal education", "1" =	"Primary school (or equivalent)", 
                                          "2" =	"Secondary/high school (or equivalent)", "3" =	"Vocational school (or equivalent)", 
                                          "4" =	"Undergraduate school/BA degree (or equivalent)", "5" =	"Graduate school/MA degree (or equivalent)", 
                                          "6" =	"PhD/MBA or similar")
levels(data_final_bl$SES_subj.t1) <- c("poorest" =	"Among the poorest", "poorer" =	"Poor", "poor" = "Below average", 
                                       "average" = "Average", 
                                       "rich" = "Above the average",  "richer" =	"Wealthy", "richest"	= "Among the wealthiest")
levels(data_final_bl$work.t1) <- c("1" = "yes, full-time", "2" = "yes, part-time", "3" = "yes, less than part time", "4" = "no")
levels(data_final_bl$study.t1) <- c("1" = "no", "2" = "yes")
levels(data_final_bl$relationship.t1) <- c("1" = "yes", "2" = "no", "3" = "other", "4" = "I don't want to tell")
levels(data_final_bl$relationship_qual.t1) <- c("1" = "married", "2" = "living together", "3" = "not living together", "4" = "I don't want to tell")
levels(data_final_bl$English_first.sc) <- c("1" = "yes", "2" = "no")
levels(data_final_bl$English_first.t1) <- c("1" = "yes", "2" = "no")
levels(data_final_bl$English_level.sc) <- c("1" = "very strong (close to native)", "2" = "strong (fluent)", "3" = "medium (I speak okay)", 
                                            "4" = "weak (I only speak a little)", "5" = "very weak" )
levels(data_final_bl$English_level.t1) <- c("1" = "very strong (close to native)", "2" = "strong (fluent)", "3" = "medium (I speak okay)", 
                                            "4" = "weak (I only speak a little)", "5" = "very weak" )
levels(data_final_bl$medication.t1) <- c("1" = "Yes", "2" = "No")
levels(data_final_bl$diagn_bpd.t1) <- c("1" = "Yes", "2" = "No")


```

```{r describe_demogr, eval=T, echo=T}
# summarise demographic variables
table_descr_demogr <- list(
  "N" = length(data_final_bl$session),
  "age mean" = mean(data_final_bl$age.sc, na.rm=T), 
  "age SD" = sd(data_final_bl$age.sc, na.rm=T),
  "gender" = table(data_final_bl$gender.sc), 
  "gender_percent" = round(prop.table(table(data_final_bl$gender.sc))*100, 2),
  "education_years_completed_percent" = round(prop.table(table(data_final_bl$edu_years_compl.t1))*100, 2),
  "education_percent" = round(prop.table(table(data_final_bl$edu_highest.t1))*100, 2), 
  "SES subjective percent" = round(prop.table(table(data_final_bl$SES_subj.t1))*100, 2), 
  "nationality" = table(data_final_bl$national.t1), 
  "work status percent" = round(prop.table(table(data_final_bl$work.t1))*100, 2),
  "study status percent" = round(prop.table(table(data_final_bl$study.t1))*100, 2),
  "relationship status percent" = round(prop.table(table(data_final_bl$relationship.t1))*100, 2),
  "relationship quality percent" = round(prop.table(table(data_final_bl$relationship_qual.t1))*100, 2),
  "English native percent" = round(prop.table(table(data_final_bl$English_first.sc))*100, 2), 
  "English level percent" = round(prop.table(table(data_final_bl$English_level.sc))*100, 2), 
  "Medication taken?" = round(prop.table(table(data_final_bl$medication.t1))*100, 2), 
  "Borderline diagnosis?" = round(prop.table(table(data_final_bl$diagn_bpd.t1))*100, 2)
  )

table_descr_demogr

# group assigment 
table(data_final_bl$Group.t1)

```

```{r BPD_conflicting_items, eval=T, echo=F}
# as per the pre-registration: "Given that the two items that compose the Absorption scale are very similarly worded in the Likert-scale response format of the instrument: 
# “I can often somehow sense the presence of another person before I actually see or hear" [bpd_absExtra_20] and 
# “At times I somehow feel the presence of someone who is not physically there” [bpd_absExtra_21]), the alternative items for the same scale are also administered in the current study, although they were initially selected to be used in the dichotomous response format version of the instrument: 
# “I sometimes “step outside” my usual self and experience an entirely different state of being.” [bpd_abs_18] and 
# “Sometimes I feel as if my mind could envelope the whole world.” [bpd_abs_19]). If the initial items correlate very highly (r>0.8) during the analyses, then the alternative items will be used for this scale for reasons of content overlap."

# check correlation
pairs.panels(data_final_bl[, c("bpd_absExtra_20.t1", "bpd_absExtra_21.t1", "bpd_abs_18.t1", "bpd_abs_19.t1")], scale=TRUE)
# therefore the initial items ("bpd_absExtra_20.t1" and "bpd_absExtra_21.t1") are kept for the scale.

```

```{r calculate_scales_variables}
## AFFECT GROUPING
data_final_bl$Group_affect <- ifelse(data_final_bl$Group.t1 %in% c("1", "2", "3"), "positive", 
                                  ifelse(data_final_bl$Group.t1 %in% c("4", "5", "6"), "neutral", 
                                         ifelse(data_final_bl$Group.t1 %in% c("7", "8", "9"), "negative", NA)))
data_final_bl$Group_affect <- factor(data_final_bl$Group_affect, levels=c("1" = "positive", "2" = "neutral", "3"="negative"))

## BORDERLINE SCREENING (MSI-BPD)
data_final_bl$BPD_screen <- rowSums(data_final_bl[, grep("^bpd_s\\d+", colnames(data_final_bl), value=T)], na.rm=F)

## MOOD:   CESD, 10 items x5 response => range: 10-50
# recode item 5 & 8 
data_final_bl[, c("cesd_5.t1", "cesd_8.t1")] <- 6 - data_final_bl[, c("cesd_5.t1", "cesd_8.t1")]
data_final_bl$CESD.t1 <-  rowSums(data_final_bl[,grep("^cesd_.+.t1", colnames(data_final_bl), value=T)], na.rm=F)

## VIDEO AFFECTS: PANAS 20 items, x7 response => range: 20-140, 10 items for positive, 10 items for negative affect
panas_positive_items <- c("panas_1.t1", "panas_3.t1", "panas_5.t1", "panas_9.t1", "panas_10.t1",
                          "panas_12.t1", "panas_14.t1", "panas_16.t1", "panas_17.t1", "panas_19.t1")
panas_negative_items <- c("panas_2.t1", "panas_4.t1", "panas_6.t1", "panas_7.t1", "panas_8.t1",
                          "panas_11.t1", "panas_13.t1", "panas_15.t1", "panas_18.t1", "panas_20.t1")

# per affect dimension (positive/negative)
data_final_bl[, "PANAS_POS.t1"] <- rowSums(data_final_bl[, panas_positive_items], na.rm = F)
data_final_bl[, "PANAS_NEG.t1"] <- rowSums(data_final_bl[, panas_negative_items], na.rm = F)

#sum score (neg-pos)
data_final_bl[, "PANAS.t1"] <- (data_final_bl[, "PANAS_NEG.t1"] - data_final_bl[, "PANAS_POS.t1"])

# PERSONALITY: IPIP 20 items, x5 response, 5 scales, each range 
bfi_items_recode <- colnames(data_final_bl[, grep("^ipip_.+[0-9]R", colnames(data_final_bl), value=F)])
data_final_bl[, bfi_items_recode] <- (6 - data_final_bl[, bfi_items_recode])
setnames(data_final_bl, names(data_final_bl[, bfi_items_recode]), gsub(pattern="R", replacement="", names(data_final_bl[, bfi_items_recode])))

# compute the scores for the items:
data_final_bl$IPIP_openness <- rowSums(data_final_bl[, grep("^ipip_agree", names(data_final_bl), value=T)], na.rm=F)
data_final_bl$IPIP_conscientiousness <- rowSums(data_final_bl[, grep("^ipip_consc", names(data_final_bl), value=T)], na.rm=F)
data_final_bl$IPIP_extraversion <- rowSums(data_final_bl[, grep("^ipip_extra", names(data_final_bl), value=T)], na.rm=F)
data_final_bl$IPIP_agreeableness <- rowSums(data_final_bl[, grep("^ipip_agree", names(data_final_bl), value=T)], na.rm=F)
data_final_bl$IPIP_neuroticism <- rowSums(data_final_bl[, grep("^ipip_neuro", names(data_final_bl), value=T)], na.rm=F)

## BORDERLINE 2 (BORNOVALOVA ET AL)
# recode coded items
bpd_recode <- grep("^bpd_.+R", colnames(data_final_bl), value=T)
data_final_bl[, bpd_recode] <- 5 - data_final_bl[, bpd_recode]
data_final_bl <- 
  data_final_bl %>% rename(bpd_contr_10.t1 = bpd_contr_10R.t1, 
               bpd_contr_11.t1 = bpd_contr_11R.t1,
               bpd_wb_16.t1 = bpd_wb_16R.t1)

data_final_bl$BPD <- rowSums(data_final_bl[, grep("^bpd_\\D{3,}", colnames(data_final_bl), value=T)], na.rm=F)
data_final_bl$BPD_stress <- rowSums(data_final_bl[, grep("^bpd_stress+", colnames(data_final_bl), value=T)], na.rm=F)
data_final_bl$BPD_alien <- rowSums(data_final_bl[, grep("^bpd_alien+", colnames(data_final_bl), value=T)], na.rm=F)
data_final_bl$BPD_control <- rowSums(data_final_bl[, grep("^bpd_contr+", colnames(data_final_bl), value=T)], na.rm=F)
data_final_bl$BPD_aggression <- rowSums(data_final_bl[, grep("^bpd_aggr+", colnames(data_final_bl), value=T)], na.rm=F)
data_final_bl$BPD_wellbeing <- rowSums(data_final_bl[, grep("^bpd_wb+", colnames(data_final_bl), value=T)], na.rm=F)
data_final_bl$BPD_abs <- rowSums(data_final_bl[, grep("^bpd_absExtra_+", colnames(data_final_bl), value=T)], na.rm=F)

## LIFE EVENT CHECKLIST
lec_items.t1 <- c(grep("lec_.+.t1$", names(data_final_bl), value=T))
data_final_bl$LEC_positive.t1 <- ifelse(data_final_bl[ , lec_items.t1] == "1" | data_final_bl[ , lec_items.t1] == "2", T, F)
data_final_bl$LEC.t1 <- rowSums(data_final_bl$LEC_positive.t1)

## EXTREMITY OF THE RESPONSE
bfi_items <- colnames(data_final_bl[, grep("^ipip_", colnames(data_final_bl), value=F)])
for (i in bfi_items) {
  data_final_bl[, paste0("extr_", i)] <- recode(data_final_bl[, i], "1" = 2, "2" = 1, "3" = 0, "4" = 1, "5" = 2)
  }

data_final_bl$EXTR <- rowSums(data_final_bl[, grep("extr_", colnames(data_final_bl), value=T)], na.rm=F)

## GENERAL IMPRESSION merge group data
gr_gen_impr_items <- c("gr1_gen_impr.t1", "gr2_gen_impr.t1", "gr3_gen_impr.t1", "gr4_gen_impr.t1", 
                       "gr5_gen_impr.t1", "gr6_gen_impr.t1", "gr7_gen_impr.t1", "gr8_gen_impr.t1", "gr9_gen_impr.t1")
data_final_bl$GEN_IMPR.t1 <- rowSums(data_final_bl[, gr_gen_impr_items], na.rm=T)  #given that every participant watched/rated one video only

  
```


```{r}
# calculate scale reliability (raw alpha)
# alpha(data_final_bl[, grep("^bpd_s\\d+", colnames(data_final_bl), value=T)])
reliability_list <- list(
  "CRONBACH ALPHAS (raw)", 
  "BPD" = round(alpha(data_final_bl[, grep("^bpd_\\D{3,}", colnames(data_final_bl), value=T)])[[1]][[1]], 2),
  "CESD" = round(alpha(data_final_bl[, grep("^cesd_\\d+", colnames(data_final_bl), value=T)])[[1]][[1]], 2),
  "PANAS positive" = round(alpha(data_final_bl[, panas_positive_items])[[1]][[1]], 2),
  "PANAS negative" = round(alpha(data_final_bl[, panas_negative_items])[[1]][[1]], 2)
  )

reliability_list



```

  



```{r describe_scales, eval=T, echo=T}
scales_continous <- c("BPD_screen", "CESD.t1", "PANAS_POS.t1", "PANAS_NEG.t1", "PANAS.t1", 
            "IPIP_openness", "IPIP_conscientiousness", "IPIP_extraversion", "IPIP_agreeableness", "IPIP_neuroticism", 
            "BPD", "BPD_stress", "BPD_alien", "BPD_control", "BPD_aggression", "BPD_abs", "EXTR", "GEN_IMPR.t1")
data_scales_continous <- data_final_bl[, scales_continous]    

descriptive_scales <- data.frame("Mean" = numeric(), "SD" = numeric(), "min" = numeric(), "max" = numeric(), "normality_SW" = numeric(), "normality_SW_p" = numeric())

for (i in 1:length(names(data_scales_continous))) {
  shapiroResult <- shapiro.test(data_scales_continous[, i])
  newRow <- list(mean(data_scales_continous[, i], na.rm=TRUE),
      sd(data_scales_continous[, i], na.rm=TRUE),
      min(data_scales_continous[, i], na.rm=TRUE),
      max(data_scales_continous[, i], na.rm=TRUE),
      shapiroResult[[1]], 
      shapiroResult[[2]])
      descriptive_scales[names(data_scales_continous[i]), ] <-  newRow
  }
print(round(descriptive_scales, 2))



```

```{r norm_plot}
par(mfrow=c(4, 5))
for (i in scales_continous) {
qqnorm(data_final_bl[, i], main = i)
}

```




```{r plots_exploratory}
hist(data_final_bl$BPD, breaks=100)

ggplot(data=data_final_bl, aes(BPD, PANAS.t1, fill=Group_affect)) +
  geom_point() + geom_jitter() + geom_smooth() +
  theme_linedraw()

ggplot(data=data_final_bl, aes(BPD, EXTR, fill=Group_affect)) +
  geom_point() +  geom_jitter() +  geom_smooth() +
  theme_linedraw()

ggplot(data=data_final_bl, aes(BPD, gen_impr_test.t1, fill=Group_affect)) +
  geom_point() +  geom_jitter() +  geom_smooth() +
  theme_linedraw()

ggplot(data=data_final_bl, aes(BPD, GEN_IMPR.t1, fill=Group_affect)) +
  geom_point() +  geom_jitter() +  geom_smooth() +
  theme_linedraw()

```


```{r save_final_data}

## save final  data
if (overwrite_data) {
  write.csv2(data_final_bl, file="Data/data_final_bl_scales_181120.csv")
}


```
