==> Scripts/Data_analysis/3_Calc_stats_LONG_Paper2.Rmd <==
---
title: "Calculated stats longitudinal Paper2"
output: html_document
---

```{r data_packages, eval=T, echo=T, message=F}
# load packages
source("../../Scripts/functions_bpd_memo.R")
packages <- c("tidyverse", "jtools", "reshape2", "ggpubr")
load_my_packages(packages)

# user-set options
overwrite = FALSE # should data/plots/tables be saved/overwritten?


# load data
data_final_long_scales <- read.csv2("../../Data/data_final_long_scales_181121.csv")

# double-check sample size
length(data_final_long_scales$session)


```



```{r visualisation_prepData, eval=T, echo=F}
# variables where mean is needed
data_viz_panas_pos <- data_final_long_scales[, c( "PANAS_POS.t1", "PANAS_POS.t2", "PANAS_POS.t3", "PANAS_POS.t4", "session")]
data_viz_panas_neg <- data_final_long_scales[, c("PANAS_NEG.t1", "PANAS_NEG.t2", "PANAS_NEG.t3", "PANAS_NEG.t4", "session")]
data_viz_cesd <-  data_final_long_scales[, c("CESD.t1", "CESD.t2", "CESD.t3", "CESD.t4", "session")]
data_viz_gen_impr <- data_final_long_scales[, c("GEN_IMPR.t1", "GEN_IMPR.t2", "GEN_IMPR.t3", "GEN_IMPR.t4", "session")]
data_viz_std <- data_final_long_scales[, c("Group_affect", "EXTR", "BPD", "session")]

# prepare the data format ggplot "likes"
library(stringr)
data_viz_panas_pos_m <- gather(data = data_viz_panas_pos, "PANAS_POS.t1", "PANAS_POS.t2", "PANAS_POS.t3", "PANAS_POS.t4", key="time", value="panas_pos")
data_viz_panas_pos_m$time <- str_sub(data_viz_panas_pos_m$time, -1, -1)

data_viz_panas_neg_m <- gather(data = data_viz_panas_neg, "PANAS_NEG.t1", "PANAS_NEG.t2", "PANAS_NEG.t3", "PANAS_NEG.t4", key="time", value="panas_neg")
data_viz_panas_neg_m$time <- str_sub(data_viz_panas_neg_m$time, -1, -1)

data_viz_cesd_m <- gather(data = data_viz_cesd, "CESD.t1", "CESD.t2", "CESD.t3", "CESD.t4", key="time", value="cesd_r")
#reverse-code mood so that higher scores indicate higher, or happier mood (currently they indicate depressed mood)
data_viz_cesd_m$cesd <- max(data_viz_cesd_m$cesd) - data_viz_cesd_m$cesd
data_viz_cesd_m$time <- str_sub(data_viz_cesd_m$time, -1, -1)

data_viz_gen_impr_m <- gather(data = data_viz_gen_impr, "GEN_IMPR.t1", "GEN_IMPR.t2", "GEN_IMPR.t3", "GEN_IMPR.t4", key="time", value="gen_impr")
data_viz_gen_impr_m$time <- str_sub(data_viz_gen_impr_m$time, -1, -1)

# merge dataframes
data_final_long_scales_viz_m <- list(data_viz_panas_pos_m, data_viz_panas_neg_m, data_viz_cesd_m, 
                                     data_viz_gen_impr_m, data_viz_std) %>% 
                                reduce(full_join)

check_gen_impr <- data_final_long_scales_viz_m[, c("time", "gen_impr", "Group_affect")]
means_gen_impr <- aggregate(check_gen_impr[, "gen_impr"], list(check_gen_impr$time, check_gen_impr$Group_affect), mean)

```

```{r save_data}
write.csv2(data_final_long_scales_viz_m, file="../../Data/data_final_long_scales_viz_m.csv")

```

```{r standardisation, eval=T}
std_vars <- c("panas_pos", "panas_neg", "cesd_r", "gen_impr")

# standardise variables
for (i in std_vars) {
  data_final_long_scales_viz_m[, i] <- scale(data_final_long_scales_viz_m[, i], center=T, scale=T)
}

# create function to calculate mean and sd
data_mean_sd_calc <- function(varname, groupnames) {
  require(plyr)
  summary_function <- function(x, col) {
    c(mean=mean(x[[col]], na.rm=TRUE), 
      sd=sd(x[[col]], na.rm = TRUE), 
      n=length(x[[col]]))
  }
  data_sum <- ddply(data_final_long_scales_viz_m, groupnames, .fun=summary_function, varname)
  data_sum <- rename(data_sum, c("mean" = paste(varname, "mean", sep="_"), "sd" = paste(varname, "sd", sep="_")))
}


```

```{r mean_sd, eval=T}
means <- c("time", "panas_pos_mean", "panas_neg_mean", "cesd_r_mean", "gen_impr_mean")
sd <- c("time", "panas_pos_sd", "panas_neg_sd", "cesd_r_sd", "gen_impr_sd")

# for the total sample
data_mean_sd_panas_pos <- data_mean_sd_calc(varname = "panas_pos", groupnames="time")
data_mean_sd_panas_neg <- data_mean_sd_calc(varname = "panas_neg", groupnames=c("time"))
data_mean_sd_cesd <- data_mean_sd_calc(varname = "cesd_r", groupnames=c("time"))
data_mean_sd_gen_impr <- data_mean_sd_calc(varname = "gen_impr", groupnames=c("time"))


data_mean_sd <- full_join(data_mean_sd_panas_pos, data_mean_sd_panas_neg, by=c("time")) %>% 
  full_join(data_mean_sd_cesd, by=c("time")) %>% 
  full_join(data_mean_sd_gen_impr, by=c("time")) 

data_plot_means_melt <- melt(data_mean_sd[, means], id="time")
rename(data_plot_means_melt, c(value="mean"))
data_plot_means_melt$variable <- str_sub(data_plot_means_melt$variable, start=1, end=-6)

data_plot_sd_melt <- melt(data_mean_sd[, sd], id="time")
rename(data_plot_sd_melt, c(value="sd"))
data_plot_sd_melt$variable <- str_sub(data_plot_sd_melt$variable, start=1, end=-4)

```



```{r join_mean_sd_n, eval=T}

# join mean and sd
data_plot_melt <- full_join(data_plot_means_melt, data_plot_sd_melt, by=c("time", "variable"))
data_plot_melt <- rename(data_plot_melt, c(value.x = "mean", value.y="sd", variable="scale"))

```


```{r mean_sd_groups, eval=T}
# for groups
means_gr <- c("time", "Group_affect", "panas_pos_mean", "panas_neg_mean", "cesd_r_mean", "gen_impr_mean")
sd_gr <- c("time",  "Group_affect", "panas_pos_sd", "panas_neg_sd", "cesd_r_sd", "gen_impr_sd")


# calculate mean and sd
data_mean_sd_panas_pos_gr <- data_mean_sd_calc(varname = "panas_pos", groupnames=c("time", "Group_affect"))
data_mean_sd_panas_neg_gr <- data_mean_sd_calc(varname = "panas_neg", groupnames=c("time", "Group_affect"))
data_mean_sd_cesd_gr <- data_mean_sd_calc(varname = "cesd_r", groupnames=c("time", "Group_affect"))
data_mean_sd_gen_impr_gr <- data_mean_sd_calc(varname = "gen_impr", groupnames=c("time", "Group_affect"))

# join data containing mean and sd
data_mean_sd_gr <- full_join(data_mean_sd_panas_pos_gr, data_mean_sd_panas_neg_gr, by=c("time", "Group_affect")) %>% 
  full_join(data_mean_sd_cesd_gr, by=c("time", "Group_affect")) %>% 
  full_join(data_mean_sd_gen_impr_gr, by=c("time", "Group_affect"))

# melt means and sd data separately, keeping the IDs (time and Group_affect)
data_plot_means_melt_gr <- melt(data_mean_sd_gr[, means_gr], id=c("time", "Group_affect"), variable_name = "scale")
data_plot_means_melt_gr <- rename(data_plot_means_melt_gr, c(value="mean"))
data_plot_means_melt_gr$scale <- str_sub(data_plot_means_melt_gr$variable, start=1, end=-6)

data_plot_sd_melt_gr <- melt(data_mean_sd_gr[, sd_gr], id=c("time", "Group_affect"), variable_name = "variable", 
                          value="sd")
data_plot_sd_melt_gr <- rename(data_plot_sd_melt_gr, c(value="sd"))
data_plot_sd_melt_gr$scale <- str_sub(data_plot_sd_melt_gr$variable, start=1, end=-4)

# join mean and sd
data_plot_melt_gr <- full_join(data_plot_means_melt_gr, data_plot_sd_melt_gr, by=c("time", "scale", "Group_affect"))



```


```{r add_group_number_to_data, eval=T}

# add group number
data_plot_melt_gr <- full_join(
  # source data with the group numbers
  aggregate(rep(1, length(data_final_long_scales_viz_m$Group_affect)),
                                 by=list(Group_affect = data_final_long_scales_viz_m$Group_affect,
                                         time = data_final_long_scales_viz_m$time), FUN=sum), 
  # data to build the plot from (summary statistics)
  data_plot_melt_gr, 
  by=c("Group_affect"="Group_affect", "time" = "time")
)

names(data_plot_melt_gr)[names(data_plot_melt_gr) == "x"] <- "n"

```

```{r visualisation_plots_faceting, eval=T, echo=T}

scale_names <- c(
  "cesd_r" = "Negative Mood \n(of the participant)",
  "gen_impr" = "Character evaluation:\nGeneral Impression",
  "panas_neg" = "Character evaluation:\nPANAS Negative",
  "panas_pos" = "Character evaluation:\nPANAS Positive")

comparisons_plot <- list(c("1", "2"), c("1", "3"), c("1", "4"), 
                         c("2", "3"), c("2", "4"), 
                         c("3", "4"))
plot_time_facet <-
  ggplot(data = data_plot_melt_gr, aes(x = time, y = mean, colour=Group_affect)) +
    geom_point(size = 2) +
    geom_line(aes(group=Group_affect), size=1.5) +
      scale_color_manual(values = c("#0072B2", "#009E73", "#D55E00")) +
                             
    # labs(title = scale) +
    ylim(-0.7, 0.8) +
  
    geom_errorbar(aes(ymin = mean - (sd / sqrt(n)), 
                      ymax = mean + (sd / sqrt(n))), 
                  width = 0.4) + 
    theme_apa(legend.pos = "top") +
    facet_grid(~ scale,
               labeller = as_labeller(scale_names)) +
    
    xlab("Time point") +
    ylab("Standardised mean value (SD)")   


plot_time_facet

ggsave("plot_time_facet_sign.png", plot=plot_time_facet, 
        width=30, height=15, units="cm")
```
