==> Scripts/Data_analysis/4_anova_longitud_Paper2_rev.Rmd <==
---
title: "4 longitudinal anova"
author: "Aniko Maraz"
date: "November 12, 2019"
output: html_document
---



```{r data_packages, eval=T, echo=T, message=F}
# load packages
source("../../Scripts/functions_bpd_memo.R")
packages <- c("tidyverse", "jtools", "reshape2",
              "ggpubr", "knitr",
              "afex", "emmeans", "Hmisc", "sjstats", 
              "multcomp" # for anova posthoc tests
              )
load_my_packages(packages)

# user-set options
overwrite = FALSE # should data/plots/tables be saved/overwritten?


# load data
data_final_long_scales_viz_m <-read.csv2("../../Data/data_final_long_scales_viz_m.csv")
data_final_long_scales_viz_m$time <- as.factor(data_final_long_scales_viz_m$time)

# double-check sample size
length(unique(data_final_long_scales_viz_m$session)) # has to be N=558
head(data_final_long_scales_viz_m)

# remove 3 missing entries (casewise) on gen_impr
data_final_long_scales_viz_m_noMissing <- data_final_long_scales_viz_m[!is.na(data_final_long_scales_viz_m$gen_impr), ]

# check variables
names(data_final_long_scales_viz_m_noMissing)

```

```{r set_variables_loops}
group_affect <- c("negative", "neutral", "positive")
time_points <- c(1, 2, 3, 4)

```


```{r ancova_center}
# center variables for ANCOVA
data_final_long_scales_viz_m_noMissing$BPD_scaled <- scale(data_final_long_scales_viz_m_noMissing$BPD, center = TRUE, scale = TRUE)
data_final_long_scales_viz_m_noMissing$cesd_scaled <- scale(data_final_long_scales_viz_m_noMissing$cesd, center = TRUE, scale = TRUE)
data_final_long_scales_viz_m_noMissing$EXTR_scaled <- scale(data_final_long_scales_viz_m_noMissing$EXTR, center = TRUE, scale = TRUE)
data_final_long_scales_viz_m_noMissing$gen_impr_scaled <- scale(data_final_long_scales_viz_m_noMissing$gen_impr, center = TRUE, scale = TRUE)
data_final_long_scales_viz_m_noMissing$panas_pos_scaled <- scale(data_final_long_scales_viz_m_noMissing$panas_pos, center = TRUE, scale = TRUE)
data_final_long_scales_viz_m_noMissing$panas_neg_scaled <- scale(data_final_long_scales_viz_m_noMissing$panas_neg, center = TRUE, scale = TRUE)

```

```{r anova_baseline, include=TRUE}
# NEGATIVELY worded evaluations
anova_base_neg <- aov_car(panas_neg ~ Group_affect + Error(X),
                          data=data_final_long_scales_viz_m_noMissing,
                          observed = c("Group_affect", "time"),
                          factorize = FALSE, type = "III", 
                          anova_table = list(correction = "HF", es ="pes"))
summary(anova_base_neg)

# check contrasts
emmeans(anova_base_neg, "Group_affect", contr = "pairwise") 



# POSITIVELY worded evaluations
anova_base_pos <- aov_car(panas_pos ~ Group_affect + Error(X),
                          data=data_final_long_scales_viz_m_noMissing,
                          observed = c("Group_affect", "time"),
                          factorize = FALSE, type = "III", 
                          anova_table = list(correction = "HF", es ="pes"))
summary(anova_base_pos)

# check contrasts
emmeans(anova_base_pos, "Group_affect", contr = "pairwise") 



# general impression
anova_base_gen_impr <- aov_car(gen_impr ~ Group_affect + Error(X),
                          data=data_final_long_scales_viz_m_noMissing,
                          observed = c("Group_affect", "time"),
                          factorize = FALSE, type = "III", 
                          anova_table = list(correction = "HF", es ="pes"))
summary(anova_base_gen_impr)

# check contrasts
emmeans(anova_base_gen_impr, "Group_affect", contr = "pairwise") 

```

```{r anova_allIn_H123, include=TRUE}
anova_base_pos_pes <- aov_car(panas_pos_scaled ~ Group_affect * time 
                       + cesd_scaled + EXTR_scaled + gen_impr_scaled + BPD_scaled + Error(X), 
                        data=data_final_long_scales_viz_m_noMissing, 
                        observed = c("Group_affect", "time"),
                        factorize = FALSE, type = "III", 
                       anova_table = list(correction = "HF", es ="pes"))

summary(anova_base_pos_pes)
summary(anova_base_pos_epsilon)
write.csv(round(as.data.frame(summary(anova_base_pos)), digits=3),
                file="H123_anova_base_pos_table.csv")


anova_base_neg <- aov_car(panas_neg_scaled ~ Group_affect * time 
                       + cesd_scaled + EXTR_scaled + gen_impr_scaled + BPD_scaled + Error(X), 
                        data=data_final_long_scales_viz_m_noMissing, 
                        observed = c("Group_affect", "time"),
                        factorize = FALSE, type = "III", 
                       anova_table = list(correction = "HF", es ="pes"))
summary(anova_base_neg)
write.csv(round(as.data.frame(summary(anova_base_neg)), digits=3),
          file="H123_anova_base_neg_table.csv")


# check contrasts
emmeans(anova_base_pos, "Group_affect", contr = "pairwise") 
emmeans(anova_base_pos, "time", contr = "pairwise") 
emmeans(anova_base_neg, "Group_affect", contr = "pairwise") 
emmeans(anova_base_neg, "time", contr = "pairwise") 

```



```{r cor_matrix, eval=T}
# check correlation between all variables in ANOVA 
rcorr(as.matrix(data_final_long_scales_viz_m_noMissing[, c("panas_pos", "panas_neg", "cesd", "gen_impr", "EXTR", "BPD")]))

```



```{r group_differences_for_plot_AND_H2.4_general_impression eval=T}
# sphericity
anova_gen_impr <- aov_ez("session", "gen_impr_scaled", between=c("Group_affect"), within = c("time"), 
                              data=data_final_long_scales_viz_m_noMissing, 
                anova_table = list(correction = "HF", es ="pes"))
summary(anova_gen_impr)

# pes
aov_car(gen_impr_scaled ~ time * Group_affect + Error(X), 
        data=data_final_long_scales_viz_m_noMissing, 
        anova_table = list(correction = "HF", es="pes"))

# pairwise comparison
summary(glht(aov(gen_impr_scaled ~ time,
                      data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$Group_affect == "negative", ]),
             linfct = mcp(time = "Tukey")), test = adjusted(type = "fdr"))


```

```{r valence_neg_H2.4, eval=T}
## does BPD correlate with the level of change in PANAS/Pos the negative group? 
# multi-level regression,, time x BPD with valence as dependent VA
multi_level_regr_genImpr <- lmer(gen_impr ~ BPD + time + (1 | session), 
                         data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$Group_affect == "negative", ])

summary(multi_level_regr_genImpr)

```

```{r group_differences_for_plot_AND_H2.5_PANAS_pos_PANAS_neg eval=T}
## PANAS negative
# sphericity
anova_gen_impr <- aov_ez("session", "panas_neg_scaled", between=c("Group_affect"), within = c("time"), 
                              data=data_final_long_scales_viz_m_noMissing, 
                anova_table = list(correction = "HF", es ="pes"))
summary(anova_gen_impr)

# pes
aov_car(panas_neg_scaled ~ time * Group_affect + Error(X), 
        data=data_final_long_scales_viz_m_noMissing, 
        anova_table = list(correction = "HF", es="pes"))

# pairwise comparison
summary(glht(aov(panas_neg_scaled ~ time,
                      data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$Group_affect == "negative", ]),
             linfct = mcp(time = "Tukey")), test = adjusted(type = "fdr"))


## PANAS positive
# sphericity
anova_gen_impr <- aov_ez("session", "panas_pos_scaled", between=c("Group_affect"), within = c("time"), 
                              data=data_final_long_scales_viz_m_noMissing, 
                anova_table = list(correction = "HF", es ="pes"))
summary(anova_gen_impr)

# pes
aov_car(panas_pos_scaled ~ time * Group_affect + Error(X), 
        data=data_final_long_scales_viz_m_noMissing, 
        anova_table = list(correction = "HF", es="pes"))

# pairwise comparison
summary(glht(aov(panas_pos_scaled ~ time,
                      data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$Group_affect == "negative", ]),
             linfct = mcp(time = "Tukey")), test = adjusted(type = "fdr"))



```



```{r valence_neg_H2.5, eval=T}
## does BPD correlate with the level of change in PANAS/Pos the negative group? 
# multi-level regression,, time x BPD with valence as dependent VA
multi_level_regr <- lmer(panas_pos_scaled ~ BPD_scaled + time + (1 | session), 
                         data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$Group_affect == "negative", ])

summary(multi_level_regr)

```

```{r variables_per_time_H2.6, eval=T}

# factors influencing valence, separate by time POSITIVE evaluations
for (i in time_points) {
  print(paste(i))
  anova <- aov(panas_pos_scaled ~ BPD_scaled + cesd_scaled + gen_impr_scaled + EXTR_scaled, 
                      data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$time == i, ])
  print(summary(anova))
  print(eta_sq(anova, partial=TRUE))
}

# factors influencing valence, separate by time NEGATIVE evaluations
for (i in time_points) {
  print(paste(i))
  anova <- aov(panas_neg_scaled ~ BPD_scaled + cesd_scaled + gen_impr_scaled + EXTR_scaled, 
                      data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$time == i, ])
  print(summary(anova))
  print(eta_sq(anova, partial=TRUE))
}

```

```{r variables_per_time_H2.6_vid_valence, eval=T}
## are the above results the same in every video valence? 

# POSITIVELY worded evaluations
for (i in time_points) {
  for (j in group_affect) {
    print(paste(i, j, sep=" and "))
    anova <- aov(panas_pos_scaled ~ BPD_scaled + cesd_scaled + gen_impr_scaled + EXTR_scaled, 
                      data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$time == i &
                                                                    data_final_long_scales_viz_m_noMissing$Group_affect == j , ])
    # print(summary(anova))
    print(eta_sq(anova, partial=TRUE))
  }
}

# NEGATIVELY worded evaluations
for (i in time_points) {
  for (j in group_affect) {
    print(paste(i, j, sep=" and "))
    anova <-aov(panas_neg_scaled ~ BPD_scaled + cesd_scaled + gen_impr_scaled + EXTR_scaled, 
                      data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$time == i &
                                                                    data_final_long_scales_viz_m_noMissing$Group_affect == j , ])
    # print(summary(anova))
    print(eta_sq(anova, partial=TRUE))
  }
}

```

```{r variables_per_time_H2.6_vid_valence_TO_TABLE, eval=T}
## are the above results the same in every video valence? 

# POSITIVELY worded evaluations
for (i in time_points) {
  for (j in group_affect) {
    print(paste(i, j, sep=" and "))
    anova <- aov(panas_pos_scaled ~ BPD_scaled + cesd_scaled + gen_impr_scaled + EXTR_scaled, 
                      data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$time == i &
                                                                    data_final_long_scales_viz_m_noMissing$Group_affect == j , ])
   print(eta_sq(anova, partial=TRUE))
  }
}

# NEGATIVELY worded evaluations
for (i in time_points) {
  for (j in group_affect) {
    print(paste(i, j, sep=" and "))
    anova <-aov(panas_neg_scaled ~ BPD_scaled + cesd_scaled + gen_impr_scaled + EXTR_scaled, 
                      data=data_final_long_scales_viz_m_noMissing[data_final_long_scales_viz_m_noMissing$time == i &
                                                                    data_final_long_scales_viz_m_noMissing$Group_affect == j , ])
    print(summary(anova))
    print(eta_sq(anova, partial=TRUE))
  }
}

```
