==> Scripts/Data_analysis/5_Calc_mediation_group_diff_BL_Paper1.Rmd <==

---
title: "Multigroup models: Mediation accross groups, comparison"
author: "Aaron Peikert"
date: "December 11, 2018"
output: html_document
---



```{r packages, eval=T, echo=F}
# load packages
library(lavaan)
library(tidyverse)
library(ggpubr)

# packages <- c("ggplot2", "tidyverse", "psych", "nortest", 
#               "apaTables", "jtools", "lavaan", "ggpubr", "semPlot") # note: semPlot (and dependencies) need a lot of memory to be installed!
# source("../functions_bpd_memo.R")
# load_my_packages(packages)

```

```{r get_data, eval=T, echo=F}
data_final_bl_scales <- read.csv2("../../Data/data_final_bl_scales_181120.csv")

```



```{r mediation, eval=T, echo=T}

model_manifest <- '
  # regressions  
    # direct effect
        PANAS_NEG.t1 ~ BPD
        PANAS_POS.t1 ~ BPD

  # mediators
        CESD.t1 ~ BPD
        EXTR ~ BPD        
        GEN_IMPR.t1 ~ BPD

        PANAS_POS.t1 ~ CESD.t1 + EXTR + GEN_IMPR.t1
        PANAS_NEG.t1 ~ CESD.t1 + EXTR + GEN_IMPR.t1

  # residual correlations
        PANAS_POS.t1 ~~ PANAS_NEG.t1

  '


# baseline modell, all parameters set free
fit_manifest_free <- sem(model_manifest, data=data_final_bl_scales, group="Group_affect")
summary(fit_manifest_free)
fitMeasures(fit_manifest_free, c("cfi", "tli", "rmsea", "srmr"))


refit <- function(model, ...){
  dots <- list(...)
  call <- as.list(model@call)
  refit <- do.call(eval(call[[1]]), c(call[-1], dots))
  refit
}

rank_models <- function(models, free, fit, desc = FALSE){
  fits <- map_dbl(models, fitmeasures, fit)
  out <- tibble(models = models, fits = fits, free = free)
  if(desc) {
    out <-  arrange(out, desc(fits))
    } else {
    out <- arrange(out, fits)
    }
  out
}

choose_free <- function(ranked_models, baseline, free, op_free){
  best <- first(ranked_models$free)
  possible <- get_parameter_names(baseline, op_free)
  possible <- setdiff(possible, best)
  free <- map(possible, ~c(.x, best))
  #browser()
  free
}

well_enough <- function(model1, model2, threshold = .05){
  if(fitmeasures(model1, "df") == fitmeasures(model2, "df"))return(TRUE)
  anova <- anova(model1, model2)
  cat(format(anova$`Pr(>Chisq)`[[2]], scientific = FALSE, digits = 8), "\n")
  isTRUE(anova$`Pr(>Chisq)`[[2]] > threshold)
}

get_parameter_names <- function(model, op){
  parameterestimates(model) %>%
    filter(op == !!op) %>%
    select(lhs, op, rhs) %>%
    pmap_chr(paste0) %>%
    unique()
}

find_differences <- function(model, restrict, op_free){
  full_restricted <- refit(model, group.equal = restrict)
  if(well_enough(model, full_restricted))return("")
  free <- get_parameter_names(model, op_free)
  find_differences_ <- function(baseline, free){
    candites <- map(free, ~refit(model, group.equal = restrict, group.partial = .x))
    ranked_models <- rank_models(candites, free, "BIC", desc = FALSE)
    if(well_enough(baseline, first(ranked_models$models)))return(ranked_models$free[[1]])
    free <- choose_free(ranked_models, baseline, free, op_free)
    if(length(free)==0)browser()
    find_differences_(baseline, free)
  }
  find_differences_(model, free)
}


find_differences(model = fit_manifest_free,
                 restrict = "regressions",
                 op_free = "~")


# baseline model, the differring paths restricted
fit_manifest_mod <- sem(model_manifest, data=data_final_bl_scales, group="Group_affect", group.equal = "regressions", group.partial = c("PANAS_POS.t1~GEN_IMPR.t1", "PANAS_NEG.t1~GEN_IMPR.t1"))

# compare the 2 models above
anova(fit_manifest_free, fit_manifest_mod)


```
